refactor: rename auth-service -> apikey-service, account-service -> kyc-service
Service names now reflect actual domain logic:
- apikey-service: API key management (issue/validate/revoke/rotate),
  scope/tier access control, TTL expiry
- kyc-service: KYC/AML processor (document verification, risk scoring,
  PEP/sanctions flagging)

Renames cover:
- services/ directories
- go.mod module paths
- Go imports (cmd/server/main.go + internal/*/test.go)
- services.yaml entries (name/path/image)
- K8s manifests (Deployment, Service)
- Dashboard ALL_SERVICES list

Verified: both services build and tests pass locally.
