--- services/account-service/go.mod
@@ -1,5 +1,0 @@
-module github.com/sinhnguyen1411/design-and-implementation-of-automated-supply-chain-security-for-go-microservices-on-kubernetes/services/account-service
-
-go 1.25.10
-
-require github.com/google/uuid v1.6.0
