+++ services/apikey-service/deploy/kubernetes/base/deployment.yaml
@@ -1,25 +1,25 @@
 piVersion: apps/v1
 kind: Deployment
 metadata:
-  name: account-service
+  name: apikey-service
   namespace: stock-trading
   labels:
-    app.kubernetes.io/name: account-service
+    app.kubernetes.io/name: apikey-service
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app.kubernetes.io/name: account-service
+      app.kubernetes.io/name: apikey-service
   template:
     metadata:
       labels:
-        app.kubernetes.io/name: account-service
+        app.kubernetes.io/name: apikey-service
       annotations:
         scs.gate/sbom-digest: ""
         scs.gate/critical-count: "0"
     spec:
       containers:
-        - name: account-service
-          image: ghcr.io/sinhnguyen1411/stock-trading/account-service
+        - name: apikey-service
+          image: ghcr.io/sinhnguyen1411/stock-trading/apikey-service
           ports:
             - containerPort: 8080
