+++ services/apikey-service/deploy/kubernetes/base/service.yaml
@@ -1,11 +1,11 @@
 piVersion: v1
 kind: Service
 metadata:
-  name: account-service
+  name: apikey-service
   namespace: stock-trading
 spec:
   selector:
-    app.kubernetes.io/name: account-service
+    app.kubernetes.io/name: apikey-service
   ports:
     - name: http
       port: 80
