+++ services/kyc-service/Dockerfile
