+++ services/kyc-service/deploy/kubernetes/base/deployment.yaml
@@ -1,25 +1,25 @@
 piVersion: apps/v1
 kind: Deployment
 metadata:
-  name: auth-service
+  name: kyc-service
   namespace: stock-trading
   labels:
-    app.kubernetes.io/name: auth-service
+    app.kubernetes.io/name: kyc-service
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app.kubernetes.io/name: auth-service
+      app.kubernetes.io/name: kyc-service
   template:
     metadata:
       labels:
-        app.kubernetes.io/name: auth-service
+        app.kubernetes.io/name: kyc-service
       annotations:
         scs.gate/sbom-digest: ""
         scs.gate/critical-count: "0"
     spec:
       containers:
-        - name: auth-service
-          image: ghcr.io/sinhnguyen1411/stock-trading/auth-service
+        - name: kyc-service
+          image: ghcr.io/sinhnguyen1411/stock-trading/kyc-service
           ports:
             - containerPort: 8080
