+++ services/kyc-service/deploy/kubernetes/base/kustomization.yaml
