+++ services/kyc-service/deploy/kubernetes/base/service.yaml
@@ -1,11 +1,11 @@
 piVersion: v1
 kind: Service
 metadata:
-  name: auth-service
+  name: kyc-service
   namespace: stock-trading
 spec:
   selector:
-    app.kubernetes.io/name: auth-service
+    app.kubernetes.io/name: kyc-service
   ports:
     - name: http
       port: 80
