+++ services/kyc-service/internal/health/health_test.go
