+++ tmp/run25810406393/security-gate-findings.json
@@ -1,0 +1,70 @@
+{
+  "findings": [
+    {
+      "cve": "CVE-2026-39820",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    },
+    {
+      "cve": "CVE-2026-42499",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    },
+    {
+      "cve": "CVE-2026-33814",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    },
+    {
+      "cve": "CVE-2026-33811",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    },
+    {
+      "cve": "CVE-2026-39836",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    },
+    {
+      "cve": "CVE-2026-42501",
+      "severity": "high",
+      "package": "stdlib",
+      "installed": "go1.25.9",
+      "fix_state": "fixed",
+      "fixed_versions": [
+        "1.25.10",
+        "1.26.3"
+      ]
+    }
+  ]
+}
